requirements.txt
diffusers
transformers
sentencepiece
accelerate
huggingface_hub
protobuf
pillow
numpy
tokenizers
charset_normalizer
Pillow
chardet
